fix logic seekbar

+++ app/src/main/java/com/example/base_widget/ui/shop/ItemCommon.kt
@@ -3,5 +3,5 @@ package com.example.base_widget.ui.shop
 data class ItemCommon(
     val image: Int,
     val name: String,
-    val time: Long
+    val time: Double
 )
